[refactor] util class 패키지 이동
