+++ src/main/java/util/StringUtils.java
@@ -1,4 +1,4 @@
-package calculator;
+package util;
 
 public class StringUtils {
